requirements.txt
discord.py
fuzzywuzzy
yt_dlp
